Refactor: keyboard 로직 분리
