+++ src/main.tsx
@@ -4,7 +4,7 @@ import { Provider } from "react-redux";
 import App from "./App";
 import "./index.css";
 import store from "./store/store";
-import setWindowHeight from "./utils/mobile";
+import setWindowHeight from "./utils/Mobile/mobile";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
 	<React.StrictMode>
